Move BundleId calculation to Manifest class
