+++ src/installer/managed/Microsoft.NET.HostModel/Bundle/Bundler.cs
@@ -9,7 +9,6 @@ using System.IO.Compression;
 using System.Linq;
 using System.Reflection.PortableExecutable;
 using System.Runtime.InteropServices;
-using System.Security.Cryptography;
 using Microsoft.NET.HostModel.AppHost;
 
 namespace Microsoft.NET.HostModel.Bundle
@@ -22,8 +21,6 @@ namespace Microsoft.NET.HostModel.Bundle
     {
         public const uint BundlerMajorVersion = 6;
         public const uint BundlerMinorVersion = 0;
-        //Same as Path.GetRandomFileName
-        private const int BundleIdLength = 12;
 
         private readonly string HostName;
         private readonly string OutputDir;
@@ -279,7 +276,6 @@ namespace Microsoft.NET.HostModel.Bundle
 
             long headerOffset = 0;
             using (BinaryWriter writer = new BinaryWriter(File.OpenWrite(bundlePath)))
-            using (SHA256 hashAlg = SHA256.Create())
             {
                 Stream bundle = writer.BaseStream;
                 bundle.Position = bundle.Length;
@@ -327,18 +323,10 @@ namespace Microsoft.NET.HostModel.Bundle
                     {
                         FileType targetType = Target.TargetSpecificFileType(type);
                         (long startOffset, long compressedSize) = AddToBundle(bundle, file, targetType);
-                        FileEntry entry = BundleManifest.AddEntry(targetType, relativePath, startOffset, file.Length, compressedSize, Target.BundleMajorVersion);
-                        file.Position = 0;
-                        byte[] hashBytes = ComputeSha256Hash(file);
-                        hashAlg.TransformBlock(hashBytes, 0, hashBytes.Length, hashBytes, 0);
+                        FileEntry entry = BundleManifest.AddEntry(targetType, file, relativePath, startOffset, compressedSize, Target.BundleMajorVersion);
                         Tracer.Log($"Embed: {entry}");
                     }
                 }
-
-
-                hashAlg.TransformFinalBlock(Array.Empty<byte>(), 0, 0);
-                byte[] manifestHash = hashAlg.Hash;
-                BundleManifest.BundleID = Convert.ToBase64String(manifestHash).Substring(BundleIdLength).Replace('/', '_');
 
                 // Write the bundle manifest
                 headerOffset = BundleManifest.Write(writer);
@@ -350,14 +338,6 @@ namespace Microsoft.NET.HostModel.Bundle
             HostWriter.SetAsBundle(bundlePath, headerOffset);
 
             return bundlePath;
-        }
-
-        private static byte[] ComputeSha256Hash(Stream stream)
-        {
-            using (SHA256 sha = SHA256.Create())
-            {
-                return sha.ComputeHash(stream);
-            }
         }
     }
 }
