+++ src/installer/managed/Microsoft.NET.HostModel/Bundle/Manifest.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Security.Cryptography;
 
 namespace Microsoft.NET.HostModel.Bundle
 {
@@ -65,7 +66,11 @@ namespace Microsoft.NET.HostModel.Bundle
         // identify this bundle. It is choosen to be compatible
         // with path-names so that the AppHost can use it in
         // extraction path.
-        public string BundleID { get; internal set; } = Path.GetRandomFileName();
+        public string BundleID => GetDeterministicId();
+        private string bundleIdInternal;
+        //Same as Path.GetRandomFileName
+        private const int BundleIdLength = 12;
+        private SHA256 hashAlg = SHA256.Create();
         public readonly uint BundleMajorVersion;
         // The Minor version is currently unused, and is always zero
         public const uint BundleMinorVersion = 0;
@@ -82,10 +87,19 @@ namespace Microsoft.NET.HostModel.Bundle
             Flags = (netcoreapp3CompatMode) ? HeaderFlags.NetcoreApp3CompatMode : HeaderFlags.None;
         }
 
-        public FileEntry AddEntry(FileType type, string relativePath, long offset, long size, long compressedSize, uint bundleMajorVersion)
+        public FileEntry AddEntry(FileType type, FileStream fileContent, string relativePath, long offset, long compressedSize, uint bundleMajorVersion)
         {
-            FileEntry entry = new FileEntry(type, relativePath, offset, size, compressedSize, bundleMajorVersion);
+            if (hashAlg == null)
+            {
+                throw new InvalidOperationException("It is forbidden to change Manifest state after it was written or BundleId was obtained.");
+            }
+
+            FileEntry entry = new FileEntry(type, relativePath, offset, fileContent.Length, compressedSize, bundleMajorVersion);
             Files.Add(entry);
+
+            fileContent.Position = 0;
+            byte[] hashBytes = ComputeSha256Hash(fileContent);
+            hashAlg.TransformBlock(hashBytes, 0, hashBytes.Length, hashBytes, 0);
 
             switch (entry.Type)
             {
@@ -104,6 +118,28 @@ namespace Microsoft.NET.HostModel.Bundle
             }
 
             return entry;
+        }
+
+        private static byte[] ComputeSha256Hash(Stream stream)
+        {
+            using (SHA256 sha = SHA256.Create())
+            {
+                return sha.ComputeHash(stream);
+            }
+        }
+
+        private string GetDeterministicId()
+        {
+            if (bundleIdInternal != null) return bundleIdInternal;
+
+            hashAlg.TransformFinalBlock(Array.Empty<byte>(), 0, 0);
+            byte[] manifestHash = hashAlg.Hash;
+            bundleIdInternal = Convert.ToBase64String(manifestHash).Substring(BundleIdLength).Replace('/', '_');
+
+            hashAlg.Dispose();
+            hashAlg = null;
+
+            return bundleIdInternal;
         }
 
         public long Write(BinaryWriter writer)
